# packages/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Movena replacement core contracts",
    "version": "2.0.0-alpha.1",
    "description": "Typed initial recovery slice. Other compatibility endpoints remain inventoried in baseline.json."
  },
  "paths": {
    "/api/v2/patient/appointments/{appointment_id}/session-notes": {
      "get": {
        "operationId": "getPatientVisitNotes",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ClinicalNoteDetail"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "appointment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
    "/api/v2/patient/today": {
      "get": {
        "operationId": "getToday",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientTodayResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
    "/api/v2/patient/consents": {
      "get": {
        "operationId": "getConsents",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ConsentRecord"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/api/v2/patient/data-rights-requests": {
      "get": {
        "operationId": "getDataRightsRequests",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DataRightsRecord"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      },
      "post": {
        "operationId": "createDataRightsRequest",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DataRightsRecord"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DataRightsCreate"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/patient/notifications": {
      "get": {
        "operationId": "getNotifications",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NotificationSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/api/v2/patient/notifications/{notification_id}/read": {
      "post": {
        "operationId": "readNotification",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotificationSummary"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "notification_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/v2/patient/reports": {
      "get": {
        "operationId": "getPatientReports",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ProgressReportSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
    "/api/v2/therapist/patients/{patient_id}/reports": {
      "post": {
        "operationId": "createProgressReport",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProgressReportCreateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "patient_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period_start",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "period_end",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "share_with_patient",
            "in": "query",
            "required": true,
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ]
      }
    },
    "/api/v2/patient/adherence": {
      "post": {
        "operationId": "saveCheckIn",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "201": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdherenceDetail"
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        },
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 128
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdherenceCreate"
              }
            }
          }
        }
      }
    },
...
    "/api/v2/patient/appointments": {
      "get": {
        "operationId": "getPatientAppointments",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AppointmentSummary"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication expired or missing"
          },
          "403": {
            "description": "Operation or object access denied"
          },
          "409": {
            "description": "Conflicting or unresolved submission"
          },
          "422": {
            "description": "Invalid input"
          },
          "503": {
            "description": "Dependency unavailable"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ClinicalNoteDetail": {
        "properties": {
          "summary": {
            "maxLength": 10000,
            "minLength": 1,
            "title": "Summary",
            "type": "string"
          },
          "recommendations": {
            "anyOf": [
              {
                "maxLength": 10000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Recommendations"
          },
          "patient_visible": {
            "default": false,
            "title": "Patient Visible",
            "type": "boolean"
          },
          "note_id": {
            "title": "Note Id",
            "type": "string"
          },
          "appointment_id": {
            "title": "Appointment Id",
            "type": "string"
          },
          "therapist_user_id": {
            "title": "Therapist User Id",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "title": "Updated At",
            "type": "string"
          },
          "author_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Author Name"
          }
        },
        "required": [
          "summary",
          "note_id",
          "appointment_id",
          "therapist_user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ClinicalNoteDetail",
        "type": "object"
      },
...
      "ConsentRecord": {
        "properties": {
          "consent_type": {
            "title": "Consent Type",
            "type": "string"
          },
          "accepted": {
            "title": "Accepted",
            "type": "boolean"
          },
          "accepted_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Accepted At"
          },
          "version": {
            "title": "Version",
            "type": "string"
          }
        },
        "required": [
          "consent_type",
          "accepted",
          "version"
        ],
        "title": "ConsentRecord",
        "type": "object"
      },
      "ProgressReportSummary": {
        "properties": {
          "progress_report_id": {
            "title": "Progress Report Id",
            "type": "string"
          },
          "period_start": {
            "format": "date",
            "title": "Period Start",
            "type": "string"
          },
          "period_end": {
            "format": "date",
            "title": "Period End",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          }
        },
        "required": [
          "progress_report_id",
          "period_start",
          "period_end",
          "created_at",
          "download_url"
        ],
        "title": "ProgressReportSummary",
        "type": "object"
      },
      "ProgressReportCreateResponse": {
        "properties": {
          "progress_report_id": {
            "title": "Progress Report Id",
            "type": "string"
          },
          "download_url": {
            "title": "Download Url",
            "type": "string"
          },
          "shared_with_patient": {
            "title": "Shared With Patient",
            "type": "boolean"
          }
        },
        "required": [
          "progress_report_id",
          "download_url",
          "shared_with_patient"
        ],
        "title": "ProgressReportCreateResponse",
        "type": "object"
      },
      "DataRightsCreate": {
        "properties": {
          "request_type": {
            "enum": [
              "export",
              "correction",
              "deletion"
            ],
            "title": "Request Type",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "maxLength": 2000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          }
        },
        "required": [
          "request_type"
        ],
        "title": "DataRightsCreate",
        "type": "object"
      },
      "DataRightsRecord": {
        "properties": {
          "request_id": {
            "title": "Request Id",
            "type": "string"
          },
          "request_type": {
            "enum": [
              "export",
              "correction",
              "deletion"
            ],
            "title": "Request Type",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Details"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Completed At"
          },
          "reviewed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Reviewed At"
          },
          "retention_until": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Retention Until"
          },
          "account_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Id"
          },
          "account_email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Email"
          },
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Account Name"
          }
        },
        "required": [
          "request_id",
          "request_type",
          "status",
          "created_at"
        ],
        "title": "DataRightsRecord",
        "type": "object"
      },
...
      "NotificationSummary": {
        "properties": {
          "notification_id": {
            "title": "Notification Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          },
          "body": {
            "title": "Body",
            "type": "string"
          },
          "action_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Action Url"
          },
          "read_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Read At"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          }
        },
        "required": [
          "notification_id",
          "kind",
          "title",
          "body",
          "created_at"
        ],
        "title": "NotificationSummary",
        "type": "object"
      },
...
      "PatientTodayResponse": {
        "properties": {
          "patient_id": {
            "title": "Patient Id",
            "type": "string"
          },
          "date": {
            "format": "date",
            "title": "Date",
            "type": "string"
          },
          "plan_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Plan Title"
          },
          "plan_items": {
            "items": {
              "$ref": "#/components/schemas/CarePlanItem"
            },
            "title": "Plan Items",
            "type": "array"
          },
          "upcoming_appointment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AppointmentSummary"
              },
              {
                "type": "null"
              }
            ],
            "default": null
          },
          "unread_notifications": {
            "default": 0,
            "title": "Unread Notifications",
            "type": "integer"
          },
          "adherence_percent_7d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Adherence Percent 7D"
          },
          "average_pain_7d": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "default": null,
            "title": "Average Pain 7D"
          },
          "completed_count_7d": {
            "default": 0,
            "title": "Completed Count 7D",
            "type": "integer"
          },
          "partial_count_7d": {
            "default": 0,
            "title": "Partial Count 7D",
            "type": "integer"
          },
          "missed_count_7d": {
            "default": 0,
            "title": "Missed Count 7D",
            "type": "integer"
          }
        },
        "required": [
          "patient_id",
          "date"
        ],
        "title": "PatientTodayResponse",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}